Add new() args files_from and argv.
The former just accumulates file names, validating only that -r is true.
The files are still actually read using files_from(), with the new()
arguments being used if no files at all are specified.

The latter specifies a reference to an array of command-line arguments
to be parsed. Options are removed from the array.