Refactor the handling of --replace output.
Instead of accumulating it and writing at the end if the file actually
changed, write it to a temp file in the input directory and rename over
the original if any changes were made.

This gave the opportunity to get rid of returning the accumulated
changes. Instead the {dry_run} argument was subverted: If it is a scalar
reference, that reference is opened for the output of the modified
content.