Dump the FLAG_PROCESS_* logic in favor of iterating.
The problem is that ack expects --ignore-ack-defaults in a file
specified by --ackrc to have effect, and the only way I can think of to
implement this is just to re-configure the object if we get to the end
and --ignore-sam-defaults is set.

I have NOT ripped out the FLAG_PROCESS_* constants. I think that is
ticklish enough to want its own commit.